# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Documentación de la API proyecto CDMI",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/eventos": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar todos los eventos",
                "description": "Devuelve una lista de todos los eventos registrados",
                "operationId": "8df3c75a189cf773a3eb105ea46e145a",
                "responses": {
                    "200": {
                        "description": "Lista de eventos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Crear un nuevo evento",
                "description": "Crea un nuevo evento en la base de datos",
                "operationId": "4594ce682158c01a68019d32f26180e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "fecha",
                                    "lugar"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Feria de Artesanías"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-01"
                                    },
                                    "lugar": {
                                        "type": "string",
                                        "example": "Centro Cultural"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/eventos/{eventoId}/registrar-usuario": {
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Registrar un usuario en un evento",
                "description": "Permite al usuario autenticado registrarse en un evento específico",
                "operationId": "1c7168c70b0c4c95294c576a23e7cebd",
                "parameters": [
                    {
                        "name": "eventoId",
                        "in": "path",
                        "description": "ID del evento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro exitoso al evento"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    },
                    "409": {
                        "description": "El usuario ya está registrado en este evento"
                    }
                }
            }
        },
...
    },
...
}